apps/ai-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
google-cloud-texttospeech==2.14.0
google-cloud-speech==2.21.0
boto3==1.34.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.0
ruff==0.1.0
black>=23.0.0